requirements.txt
neat-python
pygame
requests